+++ params.h
@@ -1,7 +1,7 @@
 // DEFINE GLOBAL PARAMETERS
 extern int nr, ntheta, nphi;
-extern double perr;
-extern double gamma1;
-extern double vGoal, vObst, vMove;
-extern double vInitial;
+extern float perr;
+extern float gamma1;
+extern float vGoal, vObst, vMove;
+extern float vInitial;
 extern int numActions;
