+++ setup.h
@@ -1,9 +1,9 @@
 // FUNCTIONS FOR SETUP
-int numberCells(double, vector<double> &);
-void setVector(int, double, vector<double> &, vector<double> &);
-void setObst(vector<double> &);
-void setGoal(vector<double> &, vector<double> &, vector<double> &);
-void conditionValue(vector<double> &, vector<double> &, vector<double> &, int, int, int);
-void setInitialValue(vector<double> &, vector<double> &, vector<double> &);
-void conditionPolicy (vector<double> &, vector<double> &, vector<char> &, int, int, int);
-void setInitialPolicy(vector<double> &, vector<double> &, vector<char> &);
+int numberCells(float, vector<float> &);
+void setVector(int, float, vector<float> &, vector<float> &);
+void setObst(vector<float> &);
+void setGoal(vector<float> &, vector<float> &, vector<float> &);
+void conditionValue(vector<float> &, vector<float> &, vector<float> &, int, int, int);
+void setInitialValue(vector<float> &, vector<float> &, vector<float> &);
+void conditionPolicy (vector<float> &, vector<float> &, vector<char> &, int, int, int);
+void setInitialPolicy(vector<float> &, vector<float> &, vector<char> &);
